+++ src/blueprint/mod.rs
@@ -1,6 +1,5 @@
 mod blueprint;
 mod compress;
-mod const_utils;
 mod from_config;
 mod into_schema;
 mod operation;
@@ -8,7 +7,6 @@ mod timeout;
 mod validate;
 
 pub use blueprint::*;
-pub use const_utils::*;
 pub use from_config::*;
 pub use operation::*;
 pub use timeout::GlobalTimeout;
